Updated hashing setup

+++ Zeeplevel/ZeeplevelService.cs
@@ -1,5 +1,5 @@
 ﻿using System.Globalization;
-using System.Numerics;
+using TNRD.Zeepkist.GTR.Backend.Numerics;
 using TNRD.Zeepkist.GTR.Backend.Zeeplevel.Resources;
 
 namespace TNRD.Zeepkist.GTR.Backend.Zeeplevel;
@@ -78,18 +78,18 @@ public class ZeeplevelService : IZeeplevelService
                 return false;
 
             level.CameraPosition = new Vector3(
-                float.Parse(splits[0], NumberStyles.Any, _culture),
-                float.Parse(splits[1], NumberStyles.Any, _culture),
-                float.Parse(splits[2], NumberStyles.Any, _culture));
+                decimal.Parse(splits[0], NumberStyles.Any, _culture),
+                decimal.Parse(splits[1], NumberStyles.Any, _culture),
+                decimal.Parse(splits[2], NumberStyles.Any, _culture));
 
             level.CameraEuler = new Vector3(
-                float.Parse(splits[3], NumberStyles.Any, _culture),
-                float.Parse(splits[4], NumberStyles.Any, _culture),
-                float.Parse(splits[5], NumberStyles.Any, _culture));
+                decimal.Parse(splits[3], NumberStyles.Any, _culture),
+                decimal.Parse(splits[4], NumberStyles.Any, _culture),
+                decimal.Parse(splits[5], NumberStyles.Any, _culture));
 
             level.CameraRotation = new Vector2(
-                float.Parse(splits[6], NumberStyles.Any, _culture),
-                float.Parse(splits[7], NumberStyles.Any, _culture));
+                decimal.Parse(splits[6], NumberStyles.Any, _culture),
+                decimal.Parse(splits[7], NumberStyles.Any, _culture));
 
             return true;
         }
@@ -155,19 +155,19 @@ public class ZeeplevelService : IZeeplevelService
                 block.Id = int.Parse(splits[0], NumberStyles.Any, _culture);
 
                 block.Position = new Vector3(
-                    float.Parse(splits[1], NumberStyles.Any, _culture),
-                    float.Parse(splits[2], NumberStyles.Any, _culture),
-                    float.Parse(splits[3], NumberStyles.Any, _culture));
+                    decimal.Parse(splits[1], NumberStyles.Any, _culture),
+                    decimal.Parse(splits[2], NumberStyles.Any, _culture),
+                    decimal.Parse(splits[3], NumberStyles.Any, _culture));
 
                 block.Euler = new Vector3(
-                    float.Parse(splits[4], NumberStyles.Any, _culture),
-                    float.Parse(splits[5], NumberStyles.Any, _culture),
-                    float.Parse(splits[6], NumberStyles.Any, _culture));
+                    decimal.Parse(splits[4], NumberStyles.Any, _culture),
+                    decimal.Parse(splits[5], NumberStyles.Any, _culture),
+                    decimal.Parse(splits[6], NumberStyles.Any, _culture));
 
                 block.Scale = new Vector3(
-                    float.Parse(splits[7], NumberStyles.Any, _culture),
-                    float.Parse(splits[8], NumberStyles.Any, _culture),
-                    float.Parse(splits[9], NumberStyles.Any, _culture));
+                    decimal.Parse(splits[7], NumberStyles.Any, _culture),
+                    decimal.Parse(splits[8], NumberStyles.Any, _culture),
+                    decimal.Parse(splits[9], NumberStyles.Any, _culture));
 
                 splits[10..27].ToList().ForEach(x => block.Paints.Add(int.Parse(x, NumberStyles.Any, _culture)));
